+++ Extensions/ListExtensions.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace Foundation.Extensions
@@ -13,7 +14,7 @@ namespace Foundation.Extensions
 		public static T Random<T>(this List<T> list)
 		{
 			if (list.Count == 0)
-				return default(T);
+				throw new System.IndexOutOfRangeException("Cannot select a random item from an empty list.");
 
 			return list[UnityEngine.Random.Range(0, list.Count)];
 		}
@@ -27,13 +28,16 @@ namespace Foundation.Extensions
 		/// <returns>Returns random object.</returns>
 		public static T Random<T>(this List<T> list, List<T> exclusions)
 		{
+			if(list.Count == 0)
+				throw new System.IndexOutOfRangeException("Cannot select a random item from an empty list.");
+
 			List<T> selectionList = new List<T>(list);
 
 			for (int i = 0; i < exclusions.Count; i++)
 				selectionList.Remove(exclusions[i]);
 
 			if(selectionList.Count == 0)
-				return default;
+				throw new System.IndexOutOfRangeException("No items left over after excluding. Cannot select a random item from an empty list.");
 
 			return selectionList[UnityEngine.Random.Range(0, selectionList.Count)];
 		}
@@ -51,6 +55,24 @@ namespace Foundation.Extensions
 
 			list.Add(item);
 			return true;
+		}
+
+		/// <summary>
+		/// Shuffle the list in place using the Fisher-Yates method.
+		/// </summary>
+		/// <param name="list">List to shuffle.</param>
+		public static void Shuffle<T>(this IList<T> list)
+		{
+			Random rng = new Random();
+			int n = list.Count;
+			while (n > 1)
+			{
+				n--;
+				int k = rng.Next(n + 1);
+				T value = list[k];
+				list[k] = list[n];
+				list[n] = value;
+			}
 		}
 	} 
 }